fix: rename emailHash anno value
